backend/pyproject.toml
[project]
name = "fiftyvoices-api"
version = "0.1.0"
description = "Fifty Voices API - FastAPI backend for voice agent platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web Framework
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "gunicorn>=23.0.0", # Production WSGI server with worker management
    # Database
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "alembic>=1.13.3",
    "psycopg2-binary>=2.9.10",
    # Redis
    "redis[hiredis]>=5.2.0",
    # Data Validation
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.1",
    "email-validator>=2.2.0",
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",
    "python-multipart>=0.0.18",
    # HTTP Client
    "httpx>=0.28.0",
    # Voice & AI
    "pipecat-ai>=0.0.67",
    "deepgram-sdk>=3.8.3",
    "elevenlabs>=1.11.0",
    "openai>=1.55.3",
    # Telephony
    "telnyx>=2.1.3",
    "twilio>=9.3.7",
    # Rate Limiting
    "slowapi>=0.1.9",
    # Monitoring & Logging
    "sentry-sdk[fastapi]>=2.18.0",
    "opentelemetry-api>=1.28.2",
    "opentelemetry-sdk>=1.28.2",
    "opentelemetry-instrumentation-fastapi>=0.49b2",
    "opentelemetry-instrumentation-sqlalchemy>=0.49b2",
    "structlog>=24.4.0",
    # WebSockets
    "websockets>=14.1",
    # Utilities
    "python-dotenv>=1.0.1",
    "pytz>=2025.2",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "faker>=33.1.0",
    "fakeredis>=2.26.2",
    "aiosqlite>=0.20.0",

    # Code Quality
    "ruff>=0.8.4",
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",

    # Type Stubs
    "types-passlib>=1.7.7.20240819",
    "types-python-jose>=3.3.4.20240106",
    "types-redis>=4.6.0.20241004",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "YTT",  # flake8-2020
    "ASYNC", # flake8-async
    "S",    # flake8-bandit (security)
    "A",    # flake8-builtins
    "COM",  # flake8-commas
    "C90",  # mccabe complexity
    "DJ",   # flake8-django
    "EM",   # flake8-errmsg
    "EXE",  # flake8-executable
    "ISC",  # flake8-implicit-str-concat
    "ICN",  # flake8-import-conventions
    "G",    # flake8-logging-format
    "PIE",  # flake8-pie
    "T20",  # flake8-print
    "PYI",  # flake8-pyi
    "PT",   # flake8-pytest-style
    "Q",    # flake8-quotes
    "RSE",  # flake8-raise
    "RET",  # flake8-return
    "SLF",  # flake8-self
    "SIM",  # flake8-simplify
    "TID",  # flake8-tidy-imports
    "TCH",  # flake8-type-checking
    "ARG",  # flake8-unused-arguments
    "PTH",  # flake8-use-pathlib
    "ERA",  # eradicate (commented code)
    "PD",   # pandas-vet
    "PL",   # pylint
    "TRY",  # tryceratops
    "NPY",  # numpy
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",    # line too long, handled by formatter
    "B008",    # do not perform function calls in argument defaults
    "C901",    # too complex
    "COM812",  # trailing comma, conflicts with formatter
    "PLR0913", # too many arguments
    "S101",    # use of assert
    "S104",    # hardcoded bind all interfaces (intentional for dev)
    "S105",    # hardcoded password (only in defaults/examples)
    "TRY003",  # long messages outside exception class
    "TRY300",  # consider else block (prefer explicit return)
    "EM101",   # raw string in exception
    "EM102",   # f-string in exception
    "RET504",  # unnecessary assignment (prefer readability)
    "PLW0603", # global statement (intentional for module-level state)
    "PLC0415", # import outside top level (conditional imports OK)
    "ARG001",  # unused function argument (FastAPI lifespan signature)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "D104"]
"tests/**/*.py" = ["S101", "PLR2004", "ARG001"]
"migrations/**/*.py" = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_unreachable = true
no_implicit_reexport = true
strict_equality = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_untyped_calls = false

[[tool.mypy.overrides]]
module = "app.db.redis"
disable_error_code = ["type-arg"]

[[tool.mypy.overrides]]
module = [
    "asyncpg.*",
    "redis.*",
    "pipecat.*",
    "deepgram.*",
    "elevenlabs.*",
    "telnyx.*",
    "twilio.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra -q --strict-markers --cov=app --cov-report=term-missing"

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/migrations/*"]

[dependency-groups]
dev = [
    "fakeredis>=2.32.1",
    "mypy>=1.18.2",
    "ruff>=0.14.6",
    "types-passlib>=1.7.7.20250602",
    "types-python-jose>=3.5.0.20250531",
]
